refactor TimerSetting modal
