+++ public/index.css
@@ -43,6 +43,7 @@ button {
 button:hover {
 	background-color: #0fb27b;
 }
+
 .button-modal {
 	width: 8vw;
 	margin: 0.4vw 0.6vw;
